Decompose action phase extras into focused panels
